+++ database/migrations/2023_08_17_090541_drop_id_user_from_reservations_table.php
@@ -12,11 +12,11 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('reservations', function (Blueprint $table) {
-            // Drop the foreign key if it exists
-            $table->dropForeign(['id_user']);  // Adjust the foreign key name if it's different
-
-            // Drop the 'id_user' column
-            $table->dropColumn('id_user');
+            // التحقق إذا كان العمود 'id_user' موجودًا قبل محاولة حذفه
+            if (Schema::hasColumn('reservations', 'id_user')) {
+                $table->dropForeign(['id_user']);
+                $table->dropColumn('id_user');
+            }
         });
     }
 
@@ -26,11 +26,7 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('reservations', function (Blueprint $table) {
-            // Recreate the 'id_user' column (assuming it's an unsigned big integer for foreign key relationship)
-            $table->unsignedBigInteger('id_user')->nullable();  // Ensure the column allows null values if needed
-
-            // Add the foreign key back, assuming it references the 'id' on the 'users' table
-            $table->foreign('id_user')->references('id')->on('users')->onDelete('cascade');
+            //
         });
     }
 };
